python-dateutil/requirements.txt
attrs==17.4.0
funcsigs==1.0.2
pluggy==0.6.0
py==1.5.2
pytest==3.3.1
python-dateutil==2.6.1
six==1.11.0
